requirements.txt
defusedxml==0.4.1
dj-config-url==0.1.1
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.10.4
django-crispy-forms==1.6.1
django-registration-redux==1.4
-e git://github.com/AASHE/django-tagulous@4f7dd6c04be0c70ccb5e23ca5652815bf3bb0258#egg=django_tagulous
gunicorn==19.6.0
html5lib==0.9999999
oauthlib==2.0.1
psycopg2==2.6.2
PyJWT==1.4.2
python-decouple==3.0
python-social-auth==0.2.21
python3-openid==3.0.10
requests==2.12.4
requests-oauthlib==0.7.0
six==1.10.0
static3==0.7.0

